test(integration): split combined plugin test into one file per plugin
CronPlugin, HttpPlugin, LoggingPlugin, and StoragePlugin each get their
own integration test file under test/bridge/integration/. Easier to run,
read, and extend with WASM counterparts later.
